Refactor sfx hook into context
