+++ app/page.js
@@ -6,12 +6,14 @@ import About from "./components/about";
 import Plaque from "./components/Plaque";
 import Settings from "./components/Settings";
 import useCustomCursor from "./hooks/useCustomCursor";
-import useSoundManager from "./hooks/useSoundEffect";
+import { useSoundContext } from "./contexts/SoundContext";
+// import useSoundManager from "./hooks/useSoundEffect";
 
 export default function Home() {
   const apiBasePath = process.env.NEXT_PUBLIC_API_BASE_PATH || "";
   const { setCursorType } = useCustomCursor(); // Access the hook
-  const { playSoundEffect } = useSoundManager();
+  const { toggleSoundEffects, toggleMusic, playSoundEffect, isSoundOn, isMusicOn } =
+    useSoundContext();
 
   const containerRef = useRef(null);
   const dragableRef = useRef(null);
@@ -65,7 +67,7 @@ export default function Home() {
   const handleShuffleBooks = useCallback(() => {
     playSoundEffect("shuffle");
     setBooks([...books].sort(() => Math.random() - 0.5)); // random shuffle
-  }, [books]);
+  }, [books, playSoundEffect]);
 
   const filteredBooks = books.filter(
     (book) =>
